02-agentic-rag/requirements.txt
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
pinecone>=5.0.0
langchain-pinecone>=0.2.0
langchain-ollama
fastapi
uvicorn[standard]
chainlit
httpx
